+++ packages/core/src/schemas/lti13/claims/baseJwtClaims.schema.ts
@@ -2,7 +2,7 @@ import * as z from 'zod';
 
 export const BaseJwtClaimsSchema = z.object({
   iss: z.string(),
-  sub: z.string().optional(),
+  sub: z.string(),
   aud: z.union([z.string(), z.array(z.string())]),
   exp: z.number(),
   iat: z.number(),